+++ gluecodium/src/test/resources/smoke/no_cache/output/android/com/example/smoke/NoCacheClass.java
@@ -1,0 +1,26 @@
+/*
+ *
+ */
+package com.example.smoke;
+import android.support.annotation.NonNull;
+import com.example.NativeBase;
+public final class NoCacheClass extends NativeBase {
+    public NoCacheClass() {
+        this(make(), (Object)null);
+    }
+    /**
+     * For internal use only.
+     * @hidden
+     */
+    protected NoCacheClass(final long nativeHandle, final Object dummy) {
+        super(nativeHandle, new Disposer() {
+            @Override
+            public void disposeNative(long handle) {
+                disposeNativeHandle(handle);
+            }
+        });
+    }
+    private static native void disposeNativeHandle(long nativeHandle);
+    private static native long make();
+    public native void foo();
+}